requirements.txt
boto3==1.9.157
Click==7.0
mysql-connector-python==8.0.16